+++ game_util_gameinfo.py
@@ -1,5 +1,2 @@
-from enum import Enum
-
-class GameInfo(Enum):
-    VERSION: str = 0.1,
-    DEBUG_MODE: bool = True
+VERSION = 0.1
+DEBUG_MODE = True